+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 
     In that case, transforming the code by inverting the condition and moving the following statements inside the branch is not valid because the function is no longer hoisted to above the branch condition. This release fixes the regression by avoiding this optimization in cases like this.
 
+* Add the option `--sourcemap=both` ([#650](https://github.com/evanw/esbuild/issues/650))
+
+    This new option puts the generated source map both an inline `//# sourceMappingURL=` data URL comment inside the output file and in an external file next to the output file. Using it is also possible with the transform API, which will cause it to return both an inline data URL comment in the `code` value and the source map JSON in the `map` value.
+
 * Tree-shake unused code with `--format=iife` ([#639](https://github.com/evanw/esbuild/issues/639))
 
     When the output format is IIFE (which wraps the code in an immediately-invoked function expression), esbuild now assumes that it's safe to remove unused code. This is an assumption that esbuild always makes when bundling but that esbuild previously didn't make when not bundling. Now esbuild will remove code even when not bundling as long as the output format is IIFE.